Put Projected_point_and_location in a file so that it gets included early enough
